requirements.txt
openai
pandas
tiktoken
scipy
python-dotenv
datasets
langchain
ragatouille
tqdm
scikit-learn
aiohttp
asyncio 